+++ src/Type/Prophet/ProphesizeDynamicReturnTypeExtension.php
@@ -13,12 +13,12 @@ declare(strict_types=1);
 
 namespace JanGregor\Prophecy\Type\Prophet;
 
-use JanGregor\Prophecy\Type\ObjectProphecy;
 use PhpParser\Node;
 use PHPStan\Analyser;
 use PHPStan\Reflection;
 use PHPStan\ShouldNotHappenException;
 use PHPStan\Type;
+use Prophecy\Prophecy;
 
 /**
  * @internal
@@ -52,7 +52,10 @@ final class ProphesizeDynamicReturnTypeExtension implements Type\DynamicMethodRe
         $returnType = $parametersAcceptor->getReturnType();
 
         if (0 === \count($methodCall->args)) {
-            return new ObjectProphecy\ObjectProphecyType();
+            return new Type\Generic\GenericObjectType(
+                Prophecy\ObjectProphecy::class,
+                []
+            );
         }
 
         $argumentType = $scope->getType($methodCall->args[0]->value);
@@ -75,6 +78,11 @@ final class ProphesizeDynamicReturnTypeExtension implements Type\DynamicMethodRe
             $className = $scope->getClassReflection()->getName();
         }
 
-        return new ObjectProphecy\ObjectProphecyType($className);
+        return new Type\Generic\GenericObjectType(
+            Prophecy\ObjectProphecy::class,
+            [
+                new Type\ObjectType($className),
+            ]
+        );
     }
 }