+++ CHANGELOG.md
@@ -8,6 +8,10 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 
 For a full diff see [`0.6.2...master`][0.6.2...master].
 
+### Changed
+
+* Moved `src/extension.neon` to `extension.neon` ([#140]), by [@localheinz]
+
 ## [`0.6.2`][0.6.2]
 
 For a full diff see [`0.6.1...0.6.2`][0.6.1...0.6.2].
@@ -118,6 +122,7 @@ For a full diff see [`afd6fd9...0.1`][afd6fd9...0.1].
 [#118]: https://github.com/Jan0707/phpstan-prophecy/pull/118
 [#119]: https://github.com/Jan0707/phpstan-prophecy/pull/119
 [#124]: https://github.com/Jan0707/phpstan-prophecy/pull/124
+[#140]: https://github.com/Jan0707/phpstan-prophecy/pull/140
 
 [@localheinz]: https://github.com/localheinz
 [@PedroTroller]: https://github.com/PedroTroller