+++ CHANGELOG.md
@@ -6,13 +6,19 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 
 ## Unreleased
 
-For a full diff see [`0.6.2...master`][0.6.2...master].
+For a full diff see [`0.7.0...master`][0.7.0...master].
+
+## [`0.7.0`][0.7.0]
+
+For a full diff see [`0.6.2...0.7.0`][0.6.2...0.7.0].
 
 ### Changed
 
 * Moved `src/extension.neon` to `extension.neon` ([#140]), by [@localheinz]
 * Marked classes as `@internal` ([#144]), by [@localheinz]
 * Moved and renamed internal classes ([#153]), by [@localheinz]
+* Required `phpstan/phpstan:~0.12.6` ([#165]), by [@Jean85]
+* Simplified extension by making use of generics ([#165]), by [@Jean85] and [@ondrejmirtes]
 
 ## [`0.6.2`][0.6.2]
 
@@ -101,6 +107,7 @@ For a full diff see [`afd6fd9...0.1`][afd6fd9...0.1].
 [0.6.0]: https://github.com/Jan0707/phpstan-prophecy/releases/tag/0.6.0
 [0.6.1]: https://github.com/Jan0707/phpstan-prophecy/releases/tag/0.6.1
 [0.6.2]: https://github.com/Jan0707/phpstan-prophecy/releases/tag/0.6.2
+[0.7.0]: https://github.com/Jan0707/phpstan-prophecy/releases/tag/0.7.0
 
 [afd6fd9...0.1]: https://github.com/Jan0707/phpstan-prophecy/compare/afd6fd9...0.1
 [0.1...0.1.1]: https://github.com/Jan0707/phpstan-prophecy/compare/0.1...0.1.1
@@ -115,7 +122,8 @@ For a full diff see [`afd6fd9...0.1`][afd6fd9...0.1].
 [0.5.1...0.6.0]: https://github.com/Jan0707/phpstan-prophecy/compare/0.5.1...0.6.0
 [0.6.0...0.6.1]: https://github.com/Jan0707/phpstan-prophecy/compare/0.6.0...0.6.1
 [0.6.1...0.6.2]: https://github.com/Jan0707/phpstan-prophecy/compare/0.6.1...0.6.2
-[0.6.2...master]: https://github.com/Jan0707/phpstan-prophecy/compare/0.6.2...master
+[0.6.2...0.7.0]: https://github.com/Jan0707/phpstan-prophecy/compare/0.6.2...0.7.0
+[0.7.0...master]: https://github.com/Jan0707/phpstan-prophecy/compare/0.7.0...master
 
 [#67]: https://github.com/Jan0707/phpstan-prophecy/pull/67
 [#79]: https://github.com/Jan0707/phpstan-prophecy/pull/79
@@ -127,6 +135,9 @@ For a full diff see [`afd6fd9...0.1`][afd6fd9...0.1].
 [#140]: https://github.com/Jan0707/phpstan-prophecy/pull/140
 [#144]: https://github.com/Jan0707/phpstan-prophecy/pull/144
 [#153]: https://github.com/Jan0707/phpstan-prophecy/pull/153
+[#165]: https://github.com/Jan0707/phpstan-prophecy/pull/165
 
+[@Jean85]: https://github.com/Jean85
 [@localheinz]: https://github.com/localheinz
+[@ondrejmirtes ]: https://github.com/ondrejmirtes
 [@PedroTroller]: https://github.com/PedroTroller